Remove flutter test pkg for now; fix integ tests
