+++ packages/amplify_flutter/example/integration_test/config_test.dart
@@ -23,7 +23,7 @@ import 'package:amplify_flutter/amplify.dart' hide Amplify;
 import 'package:amplify_flutter/src/amplify_impl.dart';
 import 'package:amplify_flutter/src/categories/amplify_categories.dart';
 import 'package:amplify_storage_s3/amplify_storage_s3.dart';
-import 'package:amplify_test_flutter/amplify_test_flutter.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
@@ -167,7 +167,8 @@ void main() {
 
   /// Calls `Amplify.reset` on the native library
   Future<void> resetAmplify() {
-    return AmplifyTestFlutter.reset();
+    const methodChannel = MethodChannel('amplify_flutter/integ_tests');
+    return methodChannel.invokeMethod<void>('reset');
   }
 
   /// Simulates a hot reload
