--- packages/amplify_test_flutter/.gitignore
@@ -1,7 +1,0 @@
-.DS_Store
-.dart_tool/
-
-.packages
-.pub/
-
-build/
