--- packages/amplify_test_flutter/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: 18116933e77adc82f80866c928266a5b4f1ed645
-  channel: stable
-
-project_type: plugin
