--- packages/amplify_test_flutter/CHANGELOG.md
@@ -1,3 +1,0 @@
-## 0.0.1
-
-* TODO: Describe initial release.
