--- packages/amplify_test_flutter/README.md
@@ -1,3 +1,0 @@
-# amplify_test_flutter
-
-Test utilities for Amplify Flutter packages
