--- packages/amplify_test_flutter/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:amplify_lints/flutter_lib.yaml
