--- packages/amplify_test_flutter/android/gradle.properties
@@ -1,3 +1,0 @@
-org.gradle.jvmargs=-Xmx1536M
-android.useAndroidX=true
-android.enableJetifier=true
