--- packages/amplify_test_flutter/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'amplify_test_flutter'
