--- packages/amplify_test_flutter/android/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.amazonaws.amplify.amplify_test_flutter">
-</manifest>
