--- packages/amplify_test_flutter/ios/Classes/AmplifyTestFlutterPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface AmplifyTestFlutterPlugin : NSObject<FlutterPlugin>
-@end
