--- packages/amplify_test_flutter/ios/Classes/IntegrationTestCase.swift
@@ -1,35 +1,0 @@
-/*
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-import Foundation
-@testable import Amplify
-import Flutter
-import XCTest
-import integration_test
-
-public struct IntegrationTestError: Error, LocalizedError {
-    public let message: String
-    
-    public var errorDescription: String? { message }
-}
-
-open class IntegrationTestCase: XCTestCase {
-    func testIntegrationTest() {
-        var testResult: NSString?
-        let integrationTestIosTest = IntegrationTestIosTest()
-        let testPass: Bool = integrationTestIosTest.testIntegrationTest(&testResult)
-        XCTAssertTrue(testPass, "\(String(describing: testResult))")
-    }
-}
