--- packages/amplify_test_flutter/ios/Classes/SwiftAmplifyTestFlutterPlugin.swift
@@ -1,36 +1,0 @@
-/*
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-import Flutter
-import UIKit
-@testable import Amplify
-
-public class SwiftAmplifyTestFlutterPlugin: NSObject, FlutterPlugin {
-  public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "amplify_test_flutter", binaryMessenger: registrar.messenger())
-    let instance = SwiftAmplifyTestFlutterPlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
-  }
-
-  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-      switch call.method {
-      case "reset":
-          Amplify.reset()
-          result(nil)
-      default:
-          result(FlutterMethodNotImplemented)
-      }
-  }
-}
